Simplify tests
