--- javascript/bin/print-meta
@@ -1,3 +1,0 @@
-#!/usr/bin/env node
-require('source-map-support').install()
-require('../dist/src/printMeta')
