+++ javascript/package.json
@@ -37,11 +37,10 @@
     "eslint-plugin-node": "11.1.0",
     "eslint-plugin-prettier": "4.0.0",
     "eslint-plugin-simple-import-sort": "7.0.0",
+    "glob": "7.2.0",
     "mocha": "9.1.3",
     "prettier": "2.5.1",
-    "source-map-support": "0.5.21",
     "ts-node": "10.4.0",
-    "tsconfig-paths": "3.12.0",
     "typescript": "4.5.2"
   },
   "directories": {
